requirements.txt
# Core PQ
liboqs-python==0.14.1
cryptography>=43

# Async / FS / system
aiofiles>=23.2.1
watchdog>=4.0.0
psutil>=5.9.8
pyyaml>=6.0.2

# Docker (préférez le plugin "docker compose" natif; éviter l'ancien paquet Python)
docker>=7.0.0
# docker-compose (legacy v1) supprimé: utilisez "docker compose" CLI

# Test
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-timeout>=2.3.0
pytest-mock>=3.14.0

# Data / science (compatibles 3.13)
numpy>=2.1.0
pandas>=2.2.2
matplotlib>=3.9.0
seaborn>=0.13.2
scipy>=1.14.0
scikit-learn>=1.5.0

# Tooling / lint / docs
jinja2>=3.1.3
black>=24.3.0
flake8>=7.0.0
pylint>=3.2.0
mypy>=1.10.0
isort>=5.13.2
sphinx>=7.3.0
sphinx-rtd-theme>=2.0.0

# Security tooling
bandit>=1.7.9
safety>=3.2.0

